store/serializers.py
from store.models import Customer
from rest_framework import serializers

class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ['id', 'user', 'birth_date',]
        read_only_fields = ['user',  ]
        
